+++ README.md
@@ -21,6 +21,7 @@ How does it work?
 * Getters and setters should access the expected fields ([`no-accessor-field-mismatch`]) ([`requires type-check`])
 * All branches in a conditional structure should not have exactly the same implementation ([`no-all-duplicated-branches`])
 * `delete` should not be used on arrays ([`no-array-delete`]) ([`requires type-check`])
+* Logical OR should not be used in switch cases ([`no-case-with-or`])
 * Collection sizes and array length comparisons should make sense ([`no-collection-size-mischeck`]) ([`requires type-check`])
 * Dead stores should be removed ([`no-dead-store`]) ([`requires type-check`])
 * Two branches in a conditional structure should not have exactly the same implementation ([`no-duplicated-branches`])
@@ -34,7 +35,6 @@ How does it work?
 * Function parameters, caught exceptions and foreach variables' initial values should not be ignored ([`no-ignored-initial-value`]) ([`requires type-check`])
 * Return values should not be ignored when function calls don't have any side effects ([`no-ignored-return`]) ([`requires type-check`])
 * Functions should use "return" consistently ([`no-inconsistent-return`])
-* Logical OR should not be used in switch cases ([`no-logical-or-in-switch-case`])
 * `Array.reverse` should not be used misleadingly ([`no-misleading-array-reverse`]) ([`requires type-check`])
 * Non-existent operators '=+', '=-' and '=!' should not be used ([`no-misspelled-operator`])
 * Multiline string literals should not be used ([`no-multiline-string-literals`])
@@ -58,6 +58,7 @@ How does it work?
 [`no-accessor-field-mismatch`]: ./sonarts-core/docs/rules/no-accessor-field-mismatch.md
 [`no-all-duplicated-branches`]: ./sonarts-core/docs/rules/no-all-duplicated-branches.md
 [`no-array-delete`]: ./sonarts-core/docs/rules/no-array-delete.md
+[`no-case-with-or`]: ./sonarts-core/docs/rules/no-case-with-or.md
 [`no-collection-size-mischeck`]: ./sonarts-core/docs/rules/no-collection-size-mischeck.md
 [`no-dead-store`]: ./sonarts-core/docs/rules/no-dead-store.md
 [`no-duplicated-branches`]: ./sonarts-core/docs/rules/no-duplicated-branches.md
@@ -71,7 +72,6 @@ How does it work?
 [`no-ignored-initial-value`]: ./sonarts-core/docs/rules/no-ignored-initial-value.md
 [`no-ignored-return`]: ./sonarts-core/docs/rules/no-ignored-return.md
 [`no-inconsistent-return`]: ./sonarts-core/docs/rules/no-inconsistent-return.md
-[`no-logical-or-in-switch-case`]: ./sonarts-core/docs/rules/no-logical-or-in-switch-case.md
 [`no-misleading-array-reverse`]: ./sonarts-core/docs/rules/no-misleading-array-reverse.md
 [`no-misspelled-operator`]: ./sonarts-core/docs/rules/no-misspelled-operator.md
 [`no-multiline-string-literals`]: ./sonarts-core/docs/rules/no-multiline-string-literals.md
