+++ sonarts-core/tests/rules/noCaseWithOrRule/noCaseWithOrRule.lint.ts
