+++ sonarts-core/tests/ruling/snapshots/no-case-with-or
