+++ sonarts-core/tslint-sonarts.json
@@ -6,6 +6,7 @@
     "no-accessor-field-mismatch": true,
     "no-all-duplicated-branches": true,
     "no-array-delete": true,
+    "no-case-with-or": true,
     "no-collection-size-mischeck": true,
     "no-dead-store": true,
     "no-duplicated-branches": true,
@@ -19,7 +20,6 @@
     "no-ignored-initial-value": true,
     "no-ignored-return": true,
     "no-inconsistent-return": false,
-    "no-logical-or-in-switch-case": true,
     "no-misleading-array-reverse": true,
     "no-misspelled-operator": true,
     "no-multiline-string-literals": true,
