+++ sonarts-sq-plugin/sonar-typescript-plugin/src/main/java/org/sonar/plugin/typescript/rules/NoCaseWithOr.java
@@ -10,17 +10,22 @@
  *
  * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
  * Lesser General Public License for more details.
  *
  * You should have received a copy of the GNU Lesser General Public License
  * along with this program; if not, write to the Free Software Foundation,
- * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+ * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-import { Rule } from "../../../src/rules/noLogicalOrInSwitchCaseRule";
-import runRule from "../../runRule";
+package org.sonar.plugin.typescript.rules;
 
-it("raises error", () => {
-  const { actualErrors, expectedErrors } = runRule(Rule, __filename);
-  expect(actualErrors).toEqual(expectedErrors);
-});
+import org.sonar.check.Rule;
+
+@Rule(key = "S3616")
+public class NoCaseWithOr extends TypeScriptRule {
+
+  @Override
+  public String tsLintKey() {
+    return "no-case-with-or";
+  }
+}
